src/ai_system.py
```python
from src.ai_deck_making import createDeck
from src.ai_networking import initiateNetworking
from src.ai_talent_help import initiateContact, scheduleInterview, postJob, matchCandidates
from src.integrations.linkedin_api import fetchLinkedInContacts
from src.integrations.job_boards_api import fetchJobBoardPosts
from src.integrations.presentation_tools_api import generatePresentation
from src.utils import validateUserProfile, validateNetworkingPreferences, validatePresentationContent, validateJobRequirements, validateCandidateProfiles

class AISystem:
    def __init__(self, user_preferences, user_profile, networking_preferences, presentation_content, presentation_structure, job_requirements, candidate_profiles):
        self.user_preferences = user_preferences
        self.user_profile = user_profile
        self.networking_preferences = networking_preferences
        self.presentation_content = presentation_content
        self.presentation_structure = presentation_structure
        self.job_requirements = job_requirements
        self.candidate_profiles = candidate_profiles

    def start(self):
        # Validate user profile
        validateUserProfile(self.user_profile)

        # Validate networking preferences
        validateNetworkingPreferences(self.networking_preferences)

        # Validate presentation content
        validatePresentationContent(self.presentation_content)

        # Validate job requirements
        validateJobRequirements(self.job_requirements)

        # Validate candidate profiles
        validateCandidateProfiles(self.candidate_profiles)

        # Start AI Deck Making
        createDeck(self.presentation_content, self.presentation_structure)

        # Start AI Networking
        initiateNetworking(self.networking_preferences)

        # Start AI Talent Help
        postJob(self.job_requirements)
        matchCandidates(self.candidate_profiles)
        initiateContact(self.candidate_profiles)
        scheduleInterview(self.candidate_profiles)
```