tests/test_ai_system.py
```python
import unittest
from src import ai_system
from src.utils import validateUserProfile, validateNetworkingPreferences, validatePresentationContent, validateJobRequirements, validateCandidateProfiles

class TestAISystem(unittest.TestCase):

    def setUp(self):
        self.ai_system = ai_system.AISystem()
        self.user_profile = {"name": "John Doe", "profession": "Software Engineer"}
        self.networking_preferences = {"industry": "Tech", "location": "San Francisco"}
        self.presentation_content = {"title": "AI in Tech", "slides": ["Introduction", "Applications", "Conclusion"]}
        self.job_requirements = {"title": "Software Engineer", "skills": ["Python", "Django"]}
        self.candidate_profiles = [{"name": "Jane Doe", "skills": ["Python", "Django"]}]

    def test_createDeck(self):
        validatePresentationContent(self.presentation_content)
        result = self.ai_system.createDeck(self.presentation_content)
        self.assertEqual(result, "deckCreationSuccess")

    def test_initiateNetworking(self):
        validateUserProfile(self.user_profile)
        validateNetworkingPreferences(self.networking_preferences)
        result = self.ai_system.initiateNetworking(self.user_profile, self.networking_preferences)
        self.assertEqual(result, "networkingInitiated")

    def test_initiateContact(self):
        validateUserProfile(self.user_profile)
        result = self.ai_system.initiateContact(self.user_profile)
        self.assertEqual(result, "contactInitiated")

    def test_scheduleInterview(self):
        validateUserProfile(self.user_profile)
        validateJobRequirements(self.job_requirements)
        validateCandidateProfiles(self.candidate_profiles)
        result = self.ai_system.scheduleInterview(self.user_profile, self.job_requirements, self.candidate_profiles)
        self.assertEqual(result, "interviewScheduled")

    def test_postJob(self):
        validateUserProfile(self.user_profile)
        validateJobRequirements(self.job_requirements)
        result = self.ai_system.postJob(self.user_profile, self.job_requirements)
        self.assertEqual(result, "jobPosted")

if __name__ == '__main__':
    unittest.main()
```